testing

+++ test_neon2.c
@@ -101,8 +101,8 @@ void main(){
         } 
     }
 
-    
-    int32_t x;
+
+    int64_t x;
 
 
     // for(k = 0; k < order; k++){
